Reorganize code for improved readability
