--- src/app/components/DeleteIcon.tsx
@@ -1,9 +1,0 @@
-import Delete from "../components/svgs/DeleteIcon";
-
-export default function DeleteIcon({ handleDeleteSession }) {
-  return (
-    <div>
-      <Delete handleDeleteSession={handleDeleteSession} />
-    </div>
-  );
-}
